// apps/cms/src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Customer {
  id: number;
  name: string;
  title?: string | null;
  slug?: string | null;
  slugLock?: boolean | null;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  description?: string | null;
  /**
   * Short description about the company
   */
  about?: string | null;
  /**
   * URL of the company website
   */
  company_url?: string | null;
  /**
   * Key statistics or metrics to highlight
   */
  stats?:
  | {
    stat: string;
    label: string;
    id?: string | null;
  }[]
  | null;
  /**
   * Miscellaneous information (e.g., Founded, Location)
   */
  misc?:
  | {
    label: string;
    text: string;
    id?: string | null;
  }[]
  | null;
  /**
   * Industry categories
   */
  industry?:
  | (
    | 'healthcare'
    | 'fintech'
    | 'ecommerce'
    | 'education'
    | 'gaming'
    | 'media'
    | 'real-estate'
    | 'saas'
    | 'social'
    | 'analytics'
    | 'ai'
    | 'developer-tools'
  )[]
  | null;
  /**
   * Size of the company
   */
  company_size?: ('startup' | 'enterprise' | 'indie_dev') | null;
  /**
   * Geographic region
   */
  region?: ('Asia' | 'Europe' | 'North America' | 'South America' | 'Africa' | 'Oceania') | null;
  /**
   * BA products being used
   */
  supabase_products?: ('database' | 'auth' | 'storage' | 'realtime' | 'functions' | 'vector')[] | null;
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
    description?: string | null;
  };
  publishedAt?: string | null;
  logo?: (number | null) | Media;
  /**
   * Light mode logo
   */
  logo_inverse?: (number | null) | Media;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}